refactor: UseCase 위치 변경
