+++ src/main/java/com/study/battleq/modules/user/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.study.battleq.modules.user.controller;
 
 import com.study.battleq.infrastructure.common.dto.ResponseDto;
 import com.study.battleq.modules.user.controller.request.LoginRequest;
-import com.study.battleq.modules.user.service.LoginUseCase;
+import com.study.battleq.modules.user.service.usecase.LoginUseCase;
 import com.study.battleq.modules.user.service.dto.TokenDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
