+++ src/main/java/com/study/battleq/modules/user/service/AuthService.java
@@ -2,6 +2,7 @@ package com.study.battleq.modules.user.service;
 
 import com.study.battleq.infrastructure.config.jwt.JwtTokenProvider;
 import com.study.battleq.modules.user.service.dto.TokenDto;
+import com.study.battleq.modules.user.service.usecase.LoginUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
